refactor: submit 이벤트를 main.ts에서 eventHandler로 이동
